Starting to Adjust EventDashboard. Removed a lot of Guest Table Code that was not being used and have started to put that into the different components. Added the GuestTable Component, Gifts Folder, EventGiftDashboard and StandaloneGiftTracker.
