+++ frontend/package.json
@@ -28,6 +28,7 @@
     "react-hot-toast": "^2.5.2",
     "react-router-dom": "^7.6.2",
     "recharts": "^2.15.3",
+    "socket.io": "^4.8.1",
     "socket.io-client": "^4.8.1"
   },
   "devDependencies": {
